move tests next to files they test
